Fixed screen not redrawing while resizing window
Our main loop starts by telling GLFW to handle queued input events. When
the window is being moved or resized, that poll will block the main thread
until the move or resize is complete. Fortunately, they give us a callback
that will fire whenever the window needs to be redrawn due to one of these
events. This commit adds a event for that callback and adds a listener in
our PostRenderSystem that tells our engine to recreate the swapchain with
the new window size, and redraw to the screen. If has to recreate the
swapchain first because otherwise the redraw will fail due to an out of
date KHR. The failure is handled by recreating the swap chain then waiting
for next frame to perform a successful draw. However, that next draw won't
happen until the new screen refresh event, which only happens when the
size of the window actually changes (and a few other situations - the
point is, not every frame). Of course, when that happens, the KHR will
probably be out of date again, causing the screen to never update until
the resize or move action is finished (generally on mouse up). So, we
recreate the swap chain before drawing. Technically that means it'll draw
when it doesn't need to in some of those aforementioned edge situations,
but it shouldn't be a big deal since these aren't *additional* renders,
these are the *only* renders that occur while mid-move/resize.

Speaking of, this'll lead to an issue later on: When resizing, the game
won't update except for the frames where the size of the window actually
changes. To remedy this, I'd suggest firing an event to pause the game
inside the framebufferResize callback (indicating the window is being
resized). That'll have to be implemented once pausing the game is
implemented, which will have to be after we start calculating deltaTime
and sending it to the systems' update methods