Replaced LuaBridge with sol2
Transitioned to being just a lua game engine. Manifest now tells engine which
world to load first, and that world defines which renderers and systems to
use, which are also defined through lua. EVERY renderer and system is
through lua. There are no core components either. Title world is used as
proof as concept. It creates a list of other worlds (NYI), has a fps
counter, and pressing the grave key toggles the imgui demo window.

There are some bugs to be worked out, and making the voxel world work, but
then this will be mergable with master